+++ gammapy/maps/core.py
@@ -13,11 +13,11 @@ import matplotlib.pyplot as plt
 import gammapy.utils.parallel as parallel
 from gammapy.utils.random import InverseCDFSampler, get_random_state
 from gammapy.utils.scripts import make_path
+from gammapy.utils.types import JsonQuantityDecoder
 from gammapy.utils.units import energy_unit_format
 from .axes import MapAxis
 from .coord import MapCoord
 from .geom import pix_tuple_to_idx
-from .io import JsonQuantityDecoder
 
 __all__ = ["Map"]
